moves engine code into the engine folder
