+++ src/engine/light.cpp
