+++ src/engine/shapegen.hpp
@@ -1,7 +1,6 @@
 #ifndef ZOSHAPEGEN_H
 #define ZOSHAPEGEN_H
 
-#include "vertex.hpp"
 #include "model.hpp"
 #include "mesh.hpp"
